Move back proper error messages to spark suite
